+++ tg_searcher/CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## dev
+### Fixed
+- Handle exception that occurs when backend init trys to find a deleted chat
+
+### Changed
+- When downloding history, now the backend will store all messages in memory and write them to index at once, to avoid blocking regular update
+
 ## [0.4.0] - 2023.2.2
 
 ### Added